Code cleanup & attempt to get applic email working
However, for some reason, when a user applies for a listing, the corresponding
email is not sent to the job's creator. If you uncomment the 'puts' line
in JobMailer's deliver_applic_email method, you'll see that this method
is indeed called and that the recipient in the "to" field is correct.
But for some reason, the email doesn't get sent!
I know the email configuration isn't wrong, because it works just fine
with the activate_job_email method in the same mailer. :/